Rename `models` to `model`
