+++ lib/cfd_trading/cfd_offer.dart
@@ -9,8 +9,8 @@ import 'package:ten_ten_one/cfd_trading/cfd_order_confirmation.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_trading.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_trading_change_notifier.dart';
 import 'package:ten_ten_one/cfd_trading/position_selection.dart';
-import 'package:ten_ten_one/models/amount.model.dart';
-import 'package:ten_ten_one/models/order.dart';
+import 'package:ten_ten_one/model/amount.model.dart';
+import 'package:ten_ten_one/model/order.dart';
 import 'package:ten_ten_one/utilities/divider.dart';
 import 'package:ten_ten_one/utilities/dropdown.dart';
 import 'package:ten_ten_one/utilities/tto_table.dart';
