+++ lib/cfd_trading/cfd_trading_change_notifier.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:ten_ten_one/models/order.dart';
+import 'package:ten_ten_one/model/order.dart';
 
 /// Responsible for managing the state across the different Cfd Trading screens.
 class CfdTradingChangeNotifier extends ChangeNotifier {
