+++ lib/main.dart
@@ -10,17 +10,17 @@ import 'package:ten_ten_one/bridge_generated/bridge_definitions.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_order_confirmation.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_order_detail.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_trading.dart';
-import 'package:ten_ten_one/models/payment.model.dart';
+import 'package:ten_ten_one/model/payment.model.dart';
 import 'package:ten_ten_one/payment_history_change_notifier.dart';
 import 'package:ten_ten_one/wallet/deposit.dart';
 import 'package:ten_ten_one/wallet/open_channel.dart';
 import 'package:ten_ten_one/wallet/wallet.dart';
 import 'package:ten_ten_one/wallet/wallet_change_notifier.dart';
-import 'package:ten_ten_one/models/amount.model.dart';
-import 'package:ten_ten_one/models/balance_model.dart';
+import 'package:ten_ten_one/model/amount.model.dart';
+import 'package:ten_ten_one/model/balance_model.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_trading_change_notifier.dart';
-import 'package:ten_ten_one/models/order.dart';
-import 'package:ten_ten_one/models/seed_backup_model.dart';
+import 'package:ten_ten_one/model/order.dart';
+import 'package:ten_ten_one/model/seed_backup_model.dart';
 import 'package:ten_ten_one/wallet/receive.dart';
 import 'package:ten_ten_one/wallet/seed.dart';
 import 'package:go_router/go_router.dart';
