+++ lib/model/balance_model.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:ten_ten_one/models/amount.model.dart';
+import 'package:ten_ten_one/model/amount.model.dart';
 
 class LightningBalance extends ChangeNotifier {
   Amount amount = Amount.zero;
